association and project querying complete. added comments and reorganized code
