+++ mern_stack/client/models/Association.js
@@ -1,15 +1,19 @@
+// model/Association.js
+
+// for linking projects to users & vice versa
+
 const mongoose = require('mongoose');
 
 // Association Schema (Many-to-Many relationship)
 const associationSchema = new mongoose.Schema({
-    user: {
-        type: mongoose.Schema.Types.ObjectId,
-        ref: 'User' // Assuming you have a User model
-    },
-    project: {
-        type: mongoose.Schema.Types.ObjectId,
-        ref: 'Project'
-    }
+	user: {
+		type: mongoose.Schema.Types.ObjectId,
+		ref: 'User'
+	},
+	project: {
+		type: mongoose.Schema.Types.ObjectId,
+		ref: 'Project'
+	}
 });
 
 const Association = mongoose.model('Association', associationSchema);
