+++ mern_stack/client/src/ErrorPage.js
@@ -1,0 +1,31 @@
+// src/ErrorPage.js
+
+// changes our page to /error if unknown page is reached
+
+import { useNavigate } from "react-router-dom";
+import React, { useEffect} from 'react';
+
+import './styles/index.css';
+
+export function Error() {
+	console.log("error page loaded");
+
+	return (
+		<div id="error-page">
+			<p>Page not found.</p>
+		</div>
+	);
+};
+
+
+const ErrorPage = () => {
+	const navigate = useNavigate();
+
+	// redirects us to the actual /error page
+	useEffect(() => {
+		navigate('/error');
+	}, []);
+	
+};
+
+export default ErrorPage;
