+++ mern_stack/client/src/Home.js
@@ -1,10 +1,13 @@
+// src/Home.js
+
+// our main navigation bar at the top
+
 import {useNavigate, Outlet} from 'react-router-dom';
-import React, {useEffect, useState} from 'react';
 import {useAuth} from './AuthContext';
 
 
 import './styles/Login.css'; // Import the custom styles
-import './index.css'
+import './styles/index.css'
 
 const Home = () => {
 	const navigate = useNavigate();
@@ -25,7 +28,6 @@ const Home = () => {
 			<div className="navigation">
 				<h1 id="home" onClick={handleHome}>Team 67</h1>
 				<h1 id="login" onClick={user ? handleProjects : handleLogin}>{user ? user : "Login"}</h1>
-				{/* <h1 id="projects" onClick={handleProjects}>{user}</h1> */}
 			</div>
 			<Outlet/>
 		</div>
