+++ mern_stack/client/src/index.js
@@ -1,26 +1,28 @@
+// index.js
+
+// the main layout of the website
+
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
+import './styles/index.css';
 import './fonts.css'
 
 import Home from './Home';
 import Login from './Login';
 import Projects from './Projects';
-import ErrorPage from './error-page';
+import ErrorPage, {Error} from './ErrorPage';
 import Register from './Register'
-
 import {AuthProvider} from './AuthContext'
-
-import reportWebVitals from './reportWebVitals';
-
 import {createBrowserRouter, RouterProvider} from 'react-router-dom';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
+
+// creates our page hierarchy
 const router = createBrowserRouter([
 	{
 		path: '/',
 		element: <Home />,
-		errorElement: <ErrorPage />,
+		errorElement: <ErrorPage />, // this is where we are sent if a page that doesn't exist is reached
 		children: [
 			{
 				path: '/login',
@@ -34,19 +36,18 @@ const router = createBrowserRouter([
 				path: '/register',
 				element: <Register />,
 			},
+			{
+				path: '/error',
+				element: <Error />
+			}
 		]
-	}
+	},
 ]);
 
 root.render(
 	<React.StrictMode>
-	  <AuthProvider>
-		<RouterProvider router={router} />
-	</AuthProvider>
+		<AuthProvider>
+			<RouterProvider router={router} />
+		</AuthProvider>
   </React.StrictMode>
 );
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
